Update LocalValidationRunner to allow intertable Relations
Update caches the metadata and molecular_metadata templates and
passes them to the validator so that it can check to ensure
relationships between tables are valid.

NOT COMPLETE
